[refactor] Move core:ktor module to core:network:ktor - Transferred all contents from core:ktor to the core:network:ktor module https://github.com/Kaaveh/ComposeNews/pull/219#discussion_r1739951830
