+++ core/network/ktor/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("composenews.android.library")
+    alias(libs.plugins.composenews.android.library)
     libs.plugins.apply {
         alias(kotlinx.serialization)
     }
