+++ core/network/ktor/src/main/java/ir/composenews/network/ApiResponseExtension.kt
@@ -1,5 +1,5 @@
 
-package ir.composenews.ktor
+package ir.composenews.network
 
 import io.ktor.client.call.body
 import io.ktor.client.statement.HttpResponse
