+++ core/network/ktor/src/main/java/ir/composenews/network/StatusCode.kt
@@ -1,5 +1,5 @@
 
-package ir.composenews.ktor
+package ir.composenews.network
 
 /**
  * https://docs.coingecko.com/reference/common-errors-rate-limit
