+++ settings.gradle
@@ -22,7 +22,7 @@ include ':core:test'
 include ':core:base'
 include ':core:uimarket'
 include ':core:extensions'
-include ':core:ktor'
+include ':core:network:ktor'
 include ':domain:market'
 include ':data:market-remote'
 include ':data:market-local'
